+++ changelog.md
@@ -4,6 +4,7 @@
 
 ## v0.1.0-beta.6
 
+- Added global `warn` for script logging at warn level.
 - Removed `:retain()` and `:release()` bindings on game objects.
 
 ## v0.1.0-beta.5